ats-resume-screener/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.8.1
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
python-multipart==0.0.9
sqlite3
nltk==3.8.1

